Include dataset and docs
Clarify the whole repo by docs with more details and dataset for
illustrations.

Moreover, a bug has been fixed, where when buttons using the
loaded picture are clicked before a image has been loaded, a
window may be declared without being shown.
Thus, if this button has been clicked twice without any loaded
image, once a image is loaded afterward, and this button is
clicked again, two windows without images and one window with the
processed loaded picture are shown.
However, only the last window is expected and wanted. Thus, the
image is checked before the window is processed now, so whether a
picture is loaded or not will be known before every thing, or at
least it should be.